requirements.txt
django==6.0.1
sqlparse==0.5.5
openpyxl==3.1.2
tzdata==2025.3
django-jazzmin==3.0.1
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.2.2
